feat: signup flow, onboarding service, email integration, seed_default_coa fix, unified agent improvements, report engine, bank sync, iron laws enforcement
- Signup: signup.py router, onboarding_service.py, email_service.py (Resend)
- Auth: middleware patch for signup bypass
- CoA: fixed seed_default_coa account types (RECEIVABLE, PAYABLE, REVENUE, COGS, OTHER_INCOME, OTHER_EXPENSE)
- Unified agent: orchestrator, tool registry, session manager, direct actions
- Reports: balance sheet, cash flow, income statement engines
- Bank sync: combination matcher, rule engine, text normalizer
- Migrations: V110-V122
- Monitoring: accounting health check

+++ backend/api_gateway/app/schemas/periods.py
@@ -7,6 +7,7 @@ Request and response models for /api/periods endpoints.
 from pydantic import BaseModel, Field, field_validator
 from typing import Optional, List
 from datetime import date, datetime
+from decimal import Decimal
 
 
 # =============================================================================
@@ -116,8 +117,8 @@ class TrialBalanceSnapshotResponse(BaseModel):
 
     id: str
     as_of_date: date
-    total_debit: float
-    total_credit: float
+    total_debit: Decimal  # Law 25: numeric(18,2)
+    total_credit: Decimal  # Law 25: numeric(18,2)
     is_balanced: bool
     generated_at: datetime
